# js/app/packages/service-clients/service-storage/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "document_storage_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/items/soup": {
      "get": {
        "tags": ["soup::inbound::axum_router"],
        "summary": "Gets the items the user has access to",
        "operationId": "get_items_soup",
        "parameters": [
          {
            "name": "expand",
            "in": "query",
            "description": "Whether to expand projects. Defaults to true.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit the number of items returned. Defaults to 20. Max 500.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "sort_method",
            "in": "query",
            "description": "Sort method. Options are viewed_at, created_at, updated_at, viewed_updated. Defaults to viewed_at.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SoupApiSort"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Base64 encoded cursor value.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoupPage"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["soup::inbound::axum_router"],
        "summary": "Gets the items the user has access to",
        "operationId": "post_items_soup",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "Base64 encoded cursor value.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSoupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoupPage"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelFilters": {
        "type": "object",
        "description": "The channel message filters used to filter down what channel messages you search over.",
        "properties": {
          "channel_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Channel IDs to search within. Examples: ['general']. Empty to search all accessible channels."
          },
          "mentions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Channel user mentions to search for. Examples: ['@username']. Empty if not filtering by mentions."
          },
          "org_id": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "Channel organization ID to search within. Empty to ignore organization filtering."
          },
          "sender_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sender IDs to search within. Examples: ['user1']. Empty to search all accessible senders."
          },
          "thread_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Channel thread IDs to search within. Examples: ['thread123']. Empty to search all threads."
          }
        }
      },
...
      "ChannelWithParticipants": {
        "type": "object",
        "required": ["channel", "participants"],
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/Channel"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelParticipant"
            }
          }
        }
      },
...
      "ChatFilters": {
        "type": "object",
        "description": "The chat filters used to filter down what chats you search over.",
        "properties": {
          "chat_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Chat ids to search over. Examples: ['chat1'], ['chat1', 'chat2']. When provided, chat search will only match results on these chats. Empty to search all accessible chats."
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Filter by chat owner. Examples: ['macro|user1@user.com'], ['macro|user1@user.com', 'macro|user2@user.com']. Empty to search all owners."
          },
          "project_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "A list of project ids to search within. Examples: ['project1']. Empty to ignore project filtering."
          },
          "role": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Chat message roles to search. Examples: ['user'], ['assistant']. Empty to search all roles."
          }
        }
      },
...
      "DocumentFilters": {
        "type": "object",
        "description": "The document filters used to filter down what documents you search over.",
        "properties": {
          "document_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Document ids to search over. Examples: ['doc1'], ['doc1', 'doc2']. Empty to search all accessible documents."
          },
          "file_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Document file types to search. Examples: ['pdf'], ['md', 'txt']. Empty to search all file types."
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Filter by document owner. Examples: ['macro|user1@user.com'], ['macro|user1@user.com', 'macro|user2@user.com']. Empty to search all owners."
          },
          "project_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "A list of project ids to search within. Examples: ['project1'].\nfiltering. Empty to ignore project filtering."
          }
        }
      },
...
      "EmailFilters": {
        "type": "object",
        "description": "The email filters used to filter down what emails you search over.",
        "properties": {
          "bcc": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Email BCC addresses to filter by. Examples: ['user@example.com']. Empty if not filtering by BCC."
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Email CC addresses to filter by. Examples: ['user@example.com']. Empty if not filtering by CC."
          },
          "recipients": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Email Recipient addresses to filter by. Examples: ['user@example.com']. Empty if not filtering by Recipient."
          },
          "senders": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Email sender addresses to filter by. Examples: ['user@example.com']. Empty to search all senders."
          }
        }
      },
...
      "EntityFilters": {
        "type": "object",
        "description": "a bundle of all of the filters for each entity type",
        "properties": {
          "channel_filters": {
            "$ref": "#/components/schemas/ChannelFilters",
            "description": "the bundled [ChannelFilters]"
          },
          "chat_filters": {
            "$ref": "#/components/schemas/ChatFilters",
            "description": "the bundled [ChatFilters]"
          },
          "document_filters": {
            "$ref": "#/components/schemas/DocumentFilters",
            "description": "the bundled [DocumentFilters]"
          },
          "email_filters": {
            "$ref": "#/components/schemas/EmailFilters",
            "description": "the bundled [EmailFilters]"
          },
          "project_filters": {
            "$ref": "#/components/schemas/ProjectFilters",
            "description": "the bundled [ProjectFilters]"
          },
          "property_filters": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PropertyFilter" },
            "description": "Property filters to apply to entities. Examples: [{\"property_id\": \"...\", \"operation\": {\"equal\": {\"values\": {\"number\": {\"values\": [5.0]}}}}}]. Empty to ignore property filtering."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "A plain old json error response for use with axum.\nyup, thats it.",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message to explain failure"
          }
        }
      },
...
      "Params": {
        "type": "object",
        "properties": {
          "expand": {
            "type": ["boolean", "null"],
            "description": "Whether to expand projects. Defaults to true."
          },
          "limit": {
            "type": ["integer", "null"],
            "format": "int32",
            "description": "Limit the number of items returned. Defaults to 20. Max 500.",
            "minimum": 0
          },
          "sort_method": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SoupApiSort",
                "description": "Sort method. Options are viewed_at, created_at, updated_at, viewed_updated. Defaults to viewed_at."
              }
            ]
          }
        }
      },
...
      "PostSoupRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EntityFilters"
          },
          {
            "$ref": "#/components/schemas/Params"
          },
          {
            "type": "object",
            "properties": {
              "emailView": {
                "type": "string",
                "description": "the view of specific emails to display"
              }
            }
          }
        ]
      },
...
      "ProjectFilters": {
        "type": "object",
        "description": "The project filters used to filter down what projects you search over.",
        "properties": {
          "owners": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Filter by project owner. Examples: ['macro|user1@user.com'], ['macro|user1@user.com', 'macro|user2@user.com']. Empty to search all owners."
          },
          "project_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Project IDs to search within. Examples: ['project1']. Empty to search all accessible projects."
          }
        }
      },
...
      "PropertyFilter": {
        "type": "object",
        "description": "Filter criteria for property values.",
        "required": ["property_id", "operation"],
        "properties": {
          "operation": {
            "$ref": "#/components/schemas/FilterOperation",
            "description": "Filter operation with embedded value(s)"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the property definition to filter by"
          }
        }
      },
...
      "SoupApiItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SoupItem"
          },
          {
            "type": "object",
            "required": ["frecency_score"],
            "properties": {
              "frecency_score": {
                "type": "number",
                "format": "double"
              }
            }
          }
        ]
      },
      "SoupApiSort": {
        "type": "string",
        "enum": [
          "viewed_at",
          "created_at",
          "updated_at",
          "viewed_updated",
          "frecency"
        ]
      },
...
      "SoupChannel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChannelWithParticipants"
          },
          {
            "$ref": "#/components/schemas/LatestMessage"
          },
          {
            "type": "object",
            "properties": {
              "interacted_at": {
                "type": ["string", "null"],
                "format": "date-time"
              },
              "viewed_at": {
                "type": ["string", "null"],
                "format": "date-time"
              }
            }
          }
        ]
      },
      "SoupChat": {
        "type": "object",
        "required": [
          "id",
          "name",
          "ownerId",
          "isPersistent",
          "createdAt",
          "updatedAt",
          "viewedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "The time the chat was created"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The chat uuid"
          },
          "isPersistent": {
            "type": "boolean",
            "description": "Whether the chat is persistent or not"
          },
          "name": {
            "type": "string",
            "description": "The name of the chat"
          },
          "ownerId": {
            "type": "string",
            "description": "Who the chat belongs to"
          },
          "projectId": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "The project id of the chat"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "description": "The time the chat was last updated"
          },
          "viewedAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "The time the chat was last viewed"
          }
        }
      },
...
      "SoupDocument": {
        "type": "object",
        "required": [
          "id",
          "documentVersionId",
          "ownerId",
          "name",
          "createdAt",
          "updatedAt",
          "viewedAt"
        ],
        "properties": {
          "branchedFromId": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "The id of the document this document branched from"
          },
          "branchedFromVersionId": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "The id of the version this document branched from\nThis could be either DocumentInstance or DocumentBom id depending on the file type"
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "The time the document was created"
          },
          "documentFamilyId": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "The id of the document family this document belongs to"
          },
          "documentVersionId": {
            "type": "integer",
            "format": "int64",
            "description": "The version of the document\nThis could be the document_instance_id or document_bom_id depending on the file type"
          },
          "fileType": {
            "type": ["string", "null"],
            "description": "The file type of the document (e.g. pdf, docx)"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The document id"
          },
          "name": {
            "type": "string",
            "description": "The name of the document"
          },
          "ownerId": {
            "type": "string",
            "description": "The owner of the document"
          },
          "projectId": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "The id of the project that this document belongs to"
          },
          "sha": {
            "type": ["string", "null"],
            "description": "If the document is a PDF, this is the SHA of the pdf\nIf the document is a DOCX, this will not be present"
          },
          "subType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SoupDocumentSubType",
                "description": "The sub type of the document if present.\nTask-related properties are encoded within the variant."
              }
            ]
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "description": "The time the document instance / document BOM was updated"
          },
          "viewedAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "The time the document was last viewed"
          }
        }
      },
      "SoupDocumentSubType": {
        "oneOf": [
          {
            "type": "object",
            "description": "A task document with its associated properties",
            "required": ["is_completed", "type"],
            "properties": {
              "is_completed": {
                "type": "boolean",
                "description": "Whether the task is completed.\nTrue if the Status property is set to \"Completed\"."
              },
              "type": {
                "type": "string",
                "enum": ["task"]
              }
            }
          }
        ],
        "description": "Sub type of a document with associated properties encoded in each variant.\nThis ensures type-safety: task properties only exist when the document is a task."
      },
      "SoupEmailThreadPreview": {
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "inboxVisible",
          "isRead",
          "isDraft",
          "isImportant",
          "sortTs",
          "createdAt",
          "updatedAt",
          "viewedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "inboxVisible": {
            "type": "boolean"
          },
          "isDraft": {
            "type": "boolean"
          },
          "isImportant": {
            "type": "boolean"
          },
          "isRead": {
            "type": "boolean"
          },
          "name": {
            "type": ["string", "null"]
          },
          "ownerId": {
            "type": "string"
          },
          "providerId": {
            "type": ["string", "null"]
          },
          "senderEmail": {
            "type": ["string", "null"]
          },
          "senderName": {
            "type": ["string", "null"]
          },
          "senderPhotoUrl": {
            "type": ["string", "null"]
          },
          "snippet": {
            "type": ["string", "null"]
          },
          "sortTs": {
            "type": "integer",
            "format": "int64"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64"
          },
          "viewedAt": {
            "type": ["integer", "null"],
            "format": "int64"
          }
        }
      },
...
      "SoupEnrichedEmailThreadPreview": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SoupEmailThreadPreview"
          },
          {
            "type": "object",
            "required": [
              "attachments",
              "attachmentsMacro",
              "participants",
              "metadata",
              "labels"
            ],
            "properties": {
              "attachments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SoupAttachment"
                }
              },
              "attachmentsMacro": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SoupMacroAttachment"
                }
              },
              "labels": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SoupLabel"
                }
              },
              "metadata": {
                "$ref": "#/components/schemas/SoupEmailThreadPreviewMetadata"
              },
              "participants": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SoupContact"
                }
              }
            }
          }
        ]
      },
      "SoupItem": {
        "oneOf": [
          {
            "type": "object",
            "required": ["data", "tag"],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SoupDocument"
              },
              "tag": {
                "type": "string",
                "enum": ["document"]
              }
            }
          },
          {
            "type": "object",
            "required": ["data", "tag"],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SoupChat"
              },
              "tag": {
                "type": "string",
                "enum": ["chat"]
              }
            }
          },
          {
            "type": "object",
            "required": ["data", "tag"],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SoupProject"
              },
              "tag": {
                "type": "string",
                "enum": ["project"]
              }
            }
          },
          {
            "type": "object",
            "required": ["data", "tag"],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SoupEnrichedEmailThreadPreview"
              },
              "tag": {
                "type": "string",
                "enum": ["emailThread"]
              }
            }
          },
          {
            "type": "object",
            "required": ["data", "tag"],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SoupChannel"
              },
              "tag": {
                "type": "string",
                "enum": ["channel"]
              }
            }
          }
        ]
      },
...
      "SoupPage": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SoupApiItem"
            }
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        }
      },
      "SoupProject": {
        "type": "object",
        "required": [
          "id",
          "name",
          "ownerId",
          "createdAt",
          "updatedAt",
          "viewedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "The time the project was created"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The id of the project"
          },
          "name": {
            "type": "string",
            "description": "The name of the project"
          },
          "ownerId": {
            "type": "string",
            "description": "The user id of who created the project"
          },
          "parentId": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "The parent project id"
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "description": "The time the project was updated"
          },
          "viewedAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "The time the document was last viewed"
          }
        }
      },
...
    }
  },
...
}