# rust/cloud-storage/soup/src/inbound/axum_router.rs
use crate::domain::{
    models::{FrecencySoupItem, SoupErr, SoupQuery, SoupRequest, SoupType},
    ports::SoupService,
};
use axum::{
    Json, Router,
    extract::{FromRef, Query, State},
    http::StatusCode,
    response::IntoResponse,
    routing::{get, post},
};
use axum_extra::{either::Either, extract::Cached};
use email::{
    domain::{
        models::{Link, PreviewView},
        ports::EmailService,
    },
    inbound::{EmailLinkErr, EmailLinkExtractor, EmailPreviewState},
};
use item_filters::{
    EntityFilters,
    ast::{EntityFilterAst, ExpandErr},
};
use macro_user_id::user_id::MacroUserIdStr;
use model_error_response::ErrorResponse;
use model_user::axum_extractor::MacroUserExtractor;
use models_pagination::{
    CursorExtractor, Frecency, PaginatedOpaqueCursor, SimpleSortMethod, SortMethod, TypeEraseCursor,
};
use models_soup::item::SoupItem;
use serde::{Deserialize, Serialize};
use std::sync::Arc;
use thiserror::Error;
use utoipa::{IntoParams, ToSchema};
use uuid::Uuid;

#[cfg(test)]
mod tests;

#[derive(Debug, Default, serde::Deserialize, IntoParams, ToSchema)]
#[into_params(parameter_in = Query)]
pub struct Params {
    /// Whether to expand projects. Defaults to true.
    #[serde(default)]
    expand: Option<bool>,
    /// Limit the number of items returned. Defaults to 20. Max 500.
    #[serde(default)]
    limit: Option<u16>,
    /// Sort method. Options are viewed_at, created_at, updated_at, viewed_updated. Defaults to viewed_at.
    #[serde(default)]
    sort_method: Option<SoupApiSort>,
}

#[derive(Debug, Deserialize, utoipa::ToSchema)]
#[serde(rename_all = "snake_case")]
pub enum SoupApiSort {
    ViewedAt,
    CreatedAt,
    UpdatedAt,
    ViewedUpdated,
    Frecency,
}

impl SoupApiSort {
    fn into_sort_method(self) -> SortMethod {
        match self {
            SoupApiSort::ViewedAt => SortMethod::Simple(SimpleSortMethod::ViewedAt),
            SoupApiSort::CreatedAt => SortMethod::Simple(SimpleSortMethod::CreatedAt),
            SoupApiSort::UpdatedAt => SortMethod::Simple(SimpleSortMethod::UpdatedAt),
            SoupApiSort::ViewedUpdated => SortMethod::Simple(SimpleSortMethod::ViewedUpdated),
            SoupApiSort::Frecency => SortMethod::Advanced(Frecency),
        }
    }
}

#[derive(Debug, Serialize, utoipa::ToSchema)]
pub struct SoupPage {
    items: Vec<SoupApiItem>,
    next_cursor: Option<String>,
}

pub struct SoupRouterState<T, U> {
    service: Arc<T>,
    email: EmailPreviewState<U>,
}

impl<T, U> Clone for SoupRouterState<T, U> {
    fn clone(&self) -> Self {
        Self {
            service: self.service.clone(),
            email: self.email.clone(),
        }
    }
}

impl<T, U> FromRef<SoupRouterState<T, U>> for EmailPreviewState<U> {
    fn from_ref(input: &SoupRouterState<T, U>) -> Self {
        input.email.clone()
    }
}

impl<T, U> SoupRouterState<T, U>
where
    T: SoupService,
    U: EmailService,
{
    pub fn new(service: T, email: U) -> Self {
        SoupRouterState {
            service: Arc::new(service),
            email: EmailPreviewState::new(email),
        }
    }

    async fn handle(
        &self,
        macro_user_id: MacroUserIdStr<'static>,
        email_link: Option<Link>,
        PostSoupRequest {
            filters,
            params,
            email_view,
        }: PostSoupRequest,
        cursor: SoupCursor,
    ) -> Result<Json<PaginatedOpaqueCursor<SoupApiItem>>, SoupHandlerErr> {
        // Extract property filters before converting to EntityFilterAst
        let property_filters = filters.property_filters.clone();

        let filters = EntityFilterAst::new_from_filters(filters)?;

        let create_fallback = move || {
            let params_sort = params
                .sort_method
                .map(|s| s.into_sort_method())
                .unwrap_or(SortMethod::Simple(SimpleSortMethod::ViewedAt));
            match params_sort {
                SortMethod::Simple(simple_sort_method) => {
                    SoupQuery::Simple(models_pagination::Query::Sort(simple_sort_method, filters))
                }
                SortMethod::Advanced(frecency) => {
                    SoupQuery::Frecency(models_pagination::Query::Sort(frecency, filters))
                }
            }
        };

        let cursor = match cursor {
            Either::E1(l) => l
                .into_option()
                .map(models_pagination::Query::Cursor)
                .map(SoupQuery::Simple)
                .unwrap_or_else(create_fallback),
            Either::E2(r) => r
                .into_option()
                .map(models_pagination::Query::Cursor)
                .map(SoupQuery::Frecency)
                .unwrap_or_else(create_fallback),
        };

        let res = self
            .service
            .get_user_soup(SoupRequest {
                soup_type: match params.expand {
                    Some(true) | None => SoupType::Expanded,
                    Some(false) => SoupType::UnExpanded,
                },
                limit: params.limit.unwrap_or(20),
                cursor,
                user: macro_user_id,
                email_preview_view: email_view,
                link_id: email_link.map(|l| l.id),
                property_filters,
            })
            .await?;

        Ok(Json(
            res.type_erase().map(SoupApiItem::from_frecency_soup_item),
        ))
    }
}

pub fn soup_router<T, U, S>(state: SoupRouterState<T, U>) -> Router<S>
where
    T: SoupService,
    U: EmailService,
    S: Send + Sync,
{
    Router::new()
        .route("/soup", get(get_soup_handler))
        .route("/soup", post(post_soup_handler))
        .with_state(state)
}

#[derive(Debug, Serialize, utoipa::ToSchema)]
pub struct SoupApiItem {
    #[serde(flatten)]
    item: SoupItem,
    frecency_score: f64,
}

impl SoupApiItem {
    fn from_frecency_soup_item(item: FrecencySoupItem) -> Self {
        let FrecencySoupItem {
            item,
            frecency_score,
        } = item;
        SoupApiItem {
            item,
            frecency_score: frecency_score
                .map(|f| f.data.frecency_score)
                .unwrap_or_default(),
        }
    }
}

#[derive(Debug, Error)]
pub enum SoupHandlerErr {
    #[error("An internal server error has occurred")]
    Internal(#[from] SoupErr),
    #[error("An internal email server error has occurred")]
    EmailLinkErr(#[from] EmailLinkErr),
    #[error("Invalid filter arguments provided")]
    ExpandErr(#[from] ExpandErr),
}

impl IntoResponse for SoupHandlerErr {
    fn into_response(self) -> axum::response::Response {
        (
            StatusCode::INTERNAL_SERVER_ERROR,
            Json(ErrorResponse {
                message: &self.to_string(),
            }),
        )
            .into_response()
    }
}

/// Gets the items the user has access to
#[utoipa::path(
    get,
    operation_id = "get_items_soup",
    path = "/items/soup",
    params(
        Params,
        ("cursor" = Option<String>, Query, description = "Base64 encoded cursor value."),
    ),
    responses(
            (status = 200, body=SoupPage),
            (status = 500, body=ErrorResponse),
    )
)]
pub async fn get_soup_handler<T, U>(
    State(service): State<SoupRouterState<T, U>>,
    Cached(MacroUserExtractor { macro_user_id, .. }): Cached<MacroUserExtractor>,
    email_link: Result<Cached<EmailLinkExtractor<U>>, EmailLinkErr>,
    Query(params): Query<Params>,
    cursor: SoupCursor,
) -> Result<Json<PaginatedOpaqueCursor<SoupApiItem>>, SoupHandlerErr>
where
    T: SoupService,
    U: EmailService,
{
    let link = match email_link {
        Ok(l) => Some(l.0.0),
        Err(EmailLinkErr::NotFound) => None,
        Err(e) => Err(e)?,
    };
    service
        .handle(
            macro_user_id,
            link,
            PostSoupRequest {
                params,
                filters: Default::default(),
                email_view: Default::default(),
            },
            cursor,
        )
        .await
}

#[derive(Debug, Deserialize, ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct PostSoupRequest {
    #[serde(default, flatten)]
    filters: EntityFilters,
    #[serde(default, flatten)]
    params: Params,
    /// the view of specific emails to display
    #[serde(default)]
    #[schema(value_type = String)]
    email_view: PreviewView,
}

type SoupCursor = axum_extra::either::Either<
    CursorExtractor<Uuid, SimpleSortMethod, Option<EntityFilterAst>>,
    CursorExtractor<Uuid, Frecency, Option<EntityFilterAst>>,
>;

/// Gets the items the user has access to
#[utoipa::path(
    post,
    operation_id = "post_items_soup",
    path = "/items/soup",
    params(
        ("cursor" = Option<String>, Query, description = "Base64 encoded cursor value."),
    ),
    responses(
            (status = 200, body=SoupPage),
            (status = 500, body=ErrorResponse),
    )
)]
#[tracing::instrument(err, skip_all)]
pub async fn post_soup_handler<T, U>(
    State(service): State<SoupRouterState<T, U>>,
    Cached(MacroUserExtractor { macro_user_id, .. }): Cached<MacroUserExtractor>,
    email_link: Result<Cached<EmailLinkExtractor<U>>, EmailLinkErr>,
    cursor: SoupCursor,
    Json(post_soup_request): Json<PostSoupRequest>,
) -> Result<Json<PaginatedOpaqueCursor<SoupApiItem>>, SoupHandlerErr>
where
    T: SoupService,
    U: EmailService,
{
    let link = match email_link {
        Ok(l) => Some(l.0.0),
        Err(EmailLinkErr::NotFound) => None,
        Err(e) => Err(e)?,
    };
    service
        .handle(macro_user_id, link, post_soup_request, cursor)
        .await
}